feat: day 11 backend

+++ src/main/java/com/example/todolist/Service/TodoService.java
@@ -2,6 +2,7 @@ package com.example.todolist.Service;
 
 import com.example.todolist.Entity.DTO.TodoRequestDTO;
 import com.example.todolist.Entity.Todo;
+import com.example.todolist.Exception.TodoNotFoundException;
 import com.example.todolist.Repository.TodoRepository;
 import org.springframework.stereotype.Service;
 
@@ -23,11 +24,29 @@ public class TodoService {
         return todoRepository.save(new Todo(todoRequestDTO.getText(),todoRequestDTO.getDone()));
     }
 
-    public Todo update(Todo todo){
+    public Todo update(Integer id, TodoRequestDTO todoRequestDTO){
+        Todo todo = todoRepository.findById(id).orElse(null);
+        if(todo == null){
+            throw new TodoNotFoundException();
+        }
+        todo.setText(todoRequestDTO.getText());
+        todo.setDone(todoRequestDTO.getDone());
         return todoRepository.save(todo);
     }
 
     public void delete(Integer id){
+        Todo todo = todoRepository.findById(id).orElse(null);
+        if(todo == null){
+            throw new TodoNotFoundException();
+        }
         todoRepository.deleteById(id);
+    }
+
+    public Todo getById(Integer id) {
+        Todo todo = todoRepository.findById(id).orElse(null);
+        if(todo == null){
+            throw new TodoNotFoundException();
+        }
+        return todo;
     }
 }
